// backend/src/types/code/purchaseCode.ts
export enum PurchaseErrorCode {
  PURCHASE_NOT_FOUND = "PURCHASE_NOT_FOUND",
  PURCHASE_FETCH_FAILED = "PURCHASE_FETCH_FAILED",
  PURCHASE_CREATE_FAILED = "PURCHASE_CREATE_FAILED",
  PURCHASE_UPDATE_FAILED = "PURCHASE_UPDATE_FAILED",
  PURCHASE_DELETE_FAILED = "PURCHASE_DELETE_FAILED",
  CARD_NOT_FOUND = "CARD_NOT_FOUND",
}